// src/service/member/schemas/getAllFollowingsParams.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Member API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */

export type GetAllFollowingsParams = {
  /**
   * 페이징 설정 페이지(0부터),한 페이지에 나올 수.
   */
  offsetWithLimit?: string;
  /**
   * 정렬 설정. 필드.정렬방법으로 요청
   */
  sorts?: string;
  /**
   * 포함시킬 하위 정보 profile:프로필
   */
  includes?: string;
};